backend/pyproject.toml
[project]
name = "gym-dm-bot"
version = "1.0.0"
description = "Instagram DM Bot + Dashboard para Gimnasio"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "pydantic-settings==2.5.2",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.2.0",
    "httpx==0.27.2",
    "openai==1.51.0",
    "pywebpush==2.0.0",
    "py-vapid==1.9.1",
    "python-multipart==0.0.12",
    "websockets==13.1",
    "tenacity==9.0.0",
]
